requirements.txt
Flask==3.0.0
cloudinary==1.36.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0